==> French Fiscal Data/notebooks/1.0 data_retrieval.ipynb <==
import os
import sys
from dotenv import load_dotenv

# Download the dataset from the S3 bucket and store it locally

s3_uri = 's3://oceandatachallengebucket/public/francetax.parquet'

output_parquet_file_path = '../data/interm/1.0_taxdata.parquet'

notebook_dir = os.path.dirname(os.path.abspath('fiscal_processing.ipynb'))
project_root = os.path.abspath(os.path.join(notebook_dir, '..'))

if project_root not in sys.path:
    sys.path.append(project_root)
    
# Load environment variables from the .env file
load_dotenv()

from datasharing.datasharing import DataSharingClient

client = DataSharingClient(duckdb_region='eu-central-1') 

# Download the dataset
client.download_dataset(s3_uri, output_parquet_file_path)
